Fixed types and interfaces
